refactor(bot): reorganize features and improve type safety in interaction handling
- Moved dice-related functions from modals to register module
- Updated imports and usage for registerDmgButton
- Enhanced type safety with Djs.GuildBasedChannel casting
- Improved code clarity by renaming embedStatistiques to registerStatistics

These changes improve code organization and maintainability.
